+++ server/src/main/java/com/main19/server/posting/like/dto/PostingLikeDto.java
@@ -1,4 +1,4 @@
-package com.main19.server.posting.dto;
+package com.main19.server.posting.like.dto;
 
 import lombok.Getter;
 import lombok.Setter;
